% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpdates4.R
\name{lpdates4}
\alias{lpdates4}
\title{Pulling Dates 2}
\usage{
lpdates2(x, year1=1981, date1="01-01", date2="12-31")
}
\arguments{
\item{x}{Date vector}

\item{year1}{The year the Date vector starts on}

\item{date1}{January 1st}

\item{date2}{December 31st}
}
\value{
A vector of indexes that correspond to the two dates every year, with respect to the start date.

NOTE: It would be significantly easier to save the output as a variable called 'dated', especially when using rasters so that you do not have to create an intermediate function.
}
\description{
This function pulls the the first and last day of every year of the date vector.
}
\examples{
dates<-seq(from = as.Date("1981-01-01"), to = as.Date("2020-12-31"), by = "day") 
lpdates4(dates, year1=1981, date1="01-01", date2="12-31")

}
