% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msdGraph.R
\name{msdGraph}
\alias{msdGraph}
\title{Mid Summer Drought Time Series Graphs}
\usage{
msdGraph(x1, x2, x3, xlon, xlat)
}
\arguments{
\item{x1}{Year 1}

\item{x2}{Year 2}

\item{x3}{Year 3}

\item{xlon}{Longitude value}

\item{xlat}{Latitude value}
}
\value{
Graph of Time Series Data
}
\description{
Plots the Time Series of Mid Summer Drought data

The input must be in the form of daily data, with the first data point being January 1st of a respective year.
}
\examples{
graph1 <- msdGraph(1981, 1985, 1986, 275, 15)

}
\seealso{
\code{\link{raster}}
}
