% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msd.R
\name{MSD}
\alias{MSD}
\title{Mid Summer Drought}
\usage{
MSD(x, dated1=dated, dated4=dated6, fun="duration")
}
\arguments{
\item{x}{RasterBrick or TimeSeries}

\item{fun}{Specify what values to be pulled from the function. Options include 'duration', 'intensity', 'max1', 'max2', 'min', and 'mindex'.}

\item{dated}{Index of Dates from the lpdates2 function (should be saved as 'dated')}

\item{dated6}{Index of Dates from the lpdates4 function (should be saved as 'dated')}
}
\value{
RasterBrick or TimeSeries of Yearly data
}
\description{
Calculate different statistics of the mid summer drought from a RasterBrick or a Time Series.

The input must be in the form of daily data, with the first data point being January 1st of a respective year.

If x is a RasterBrick, then the output is a RasterBrick with a data point for each year. If you would like to find a statistical value of the new rasterbrick, it is recommended to use

r <- raster::calc(x, mean)

If the index vector from lpdates2/lpdates4 is not saved as 'dated'/'dated6', a function will have to be created to use it in raster::calc.
}
\examples{
# using RasterBricks
r<-raster::calc(x, MSD)

# using TimeSeries
r<-MSD(x)
}
\seealso{
\code{\link{raster}}
}
